pyproject.toml
[project]
name = "osmose-presets"
license = {file = "LICENSE.txt"}
version = "0.2.6"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Carl Caulkett", email = "carl.caulkett@gmail.com" },
]
requires-python = ">=3.10"
dependencies = [
    "mido>=1.3.3",
    "python-rtmidi>=1.5.8",
    "textual[syntax]>=6.1.0",
]

[project.scripts]
osmose-presets = "osmose_presets.app:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "textual-dev>=1.7.0",
]

[tool.ruff]
indent-width = 3
line-length = 160
target-version = "py313"
